layers: Add dedicated YCbCr validation file
